Move functionality into TextSourceGenerator
