+++ ext/js/data/sandbox/string-util.js
@@ -68,3 +68,11 @@ export function readCodePointsBackward(text, position, count) {
     }
     return result;
 }
+
+/**
+ * @param {string} string
+ * @returns {boolean}
+ */
+export function isWhitespace(string) {
+    return string.trim().length === 0;
+}
